Fixed jquery library conflicts
The second jQuery library was being injected before the title tag, so I
moved my script includes below the title tag and took out the jquery
include that I used so that the only one that loads up is theirs and my
files load in after that. I believe the reason for your jquery library
was the country picklist, so I tested it and it works. Shouldn't have
any more JS bugs, let me know if there are issues.